renames analysis.py to indicators.py
